WIP: reducer positions -> moves
